graph_core/requirements.txt
grpcio==1.29.0
grpcio-tools==1.29.0
protobuf==3.12.2
pybuilder==0.12.6
sqlalchemy==1.3.18
snowflake-sqlalchemy==1.2.3
snowflake-connector-python==2.3.2
requests==2.23.0
python-dateutil==2.8.1
pandas==1.1.2
pyarrow==1.0.1
numpy==1.19.2
multipledispatch==0.6.0
azure-storage-blob==12.3.2
azure-core==1.6.0
abcplus==0.1.0